% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_ParametersSimulation.R
\name{Alternator}
\alias{Alternator}
\title{Migration}
\usage{
Alternator(P, S, repeats)
}
\arguments{
\item{P}{A list of parameters.}

\item{S}{A list of data structures.}

\item{repeats}{how many times to repeat migration.}
}
\description{
A function wrapper that alternates between migration and horizontal transfer.
}
\keyword{SimParam}
