% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenPopulation.R
\name{Extinction}
\alias{Extinction}
\title{Extinction}
\usage{
Extinction(populations, occupied)
}
\arguments{
\item{populations}{The data for all exising populations.}

\item{occupied}{The territories with a population on them.}
}
\description{
Tests which populations will die based on population size and random chance.
}
\keyword{Population_Dynamics}
