% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Phonemes.R
\name{GeneratePhonemeProbabilities}
\alias{GeneratePhonemeProbabilities}
\title{Generate Phoneme Probabilities}
\usage{
GeneratePhonemeProbabilities(P)
}
\arguments{
\item{P}{A list of parameters.}
}
\description{
Genetrates a vector of the probability to know each phoneme.
}
\keyword{Phonemes}
