% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenPopulation.R
\name{MakePopulation}
\alias{MakePopulation}
\title{Make Population}
\usage{
MakePopulation(P, population)
}
\arguments{
\item{P}{A list of parameters.}

\item{population}{The population data used to make a new population.}
}
\description{
Generates new population based on the parent population.
}
\keyword{Population_Dynamics}
