% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenPopulation.R
\name{PopulationGrowth}
\alias{PopulationGrowth}
\title{Population Growth}
\usage{
PopulationGrowth(P, populationSizes, occupied)
}
\arguments{
\item{P}{A list of parameters.}

\item{populationSizes}{The number of people live on each territory.}

\item{occupied}{The indicies of territories with people living on them.}
}
\description{
Adds new individuals to existing populations when population size is used in the simulation.
}
\keyword{Population_Dynamics}
