% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Plotting.R
\name{PopulationPlot}
\alias{PopulationPlot}
\title{Population Plot}
\usage{
PopulationPlot(P, Data, groups = NA, colors = NA)
}
\arguments{
\item{P}{A list of parameters.}

\item{groups}{Group structure of which territories were descended from what population seed.}

\item{colors}{A vector of colors of length equal to the number of seed populations.}
}
\description{
Population Plot
}
\keyword{Plotting}
