% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Directions.R
\name{StepDirections}
\alias{StepDirections}
\title{Step Directions}
\usage{
StepDirections(P)
}
\arguments{
\item{P}{A list of parameters.}
}
\description{
A wrapper that calls StepOne(), add barriers if required, then expands StepOne as many steps as the Steps parameter calls for.
}
\keyword{Directions}
\keyword{Wrapper}
