% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_GenLanguage.R
\name{UpdateExistingPhonemes}
\alias{UpdateExistingPhonemes}
\title{Update Existing Phonemes}
\usage{
UpdateExistingPhonemes(existingPhonemes, newMut, index)
}
\arguments{
\item{newMut}{The new mutation generated by GetAMutation().}

\item{index}{Whether to enact changes to the langauge based on the first or section member of the mutation structure.}

\item{ExistingPhonemes}{The Phonemes currently in the language.}
}
\description{
Change the language to incorporation new mutations.
}
\keyword{Language}
